DRY the add photo buttons

+++ app/helpers/buttons_helper.rb
@@ -1,12 +1,4 @@
 module ButtonsHelper
-  def seed_add_photo_button(seed)
-    button(new_photo_path(id: seed.id, type: "seed"), 'buttons.add_a_photo', 'camera')
-  end
-  
-  def harvest_add_photo_button(harvest)
-    button(new_photo_path(id: harvest.id, type: "harvest"), 'buttons.add_a_photo', 'camera')
-  end
-
   def garden_plant_something_button(garden)
     button(new_planting_path(garden_id: garden.id), 'buttons.plant_something_here', 'leaf')
   end
@@ -24,14 +16,10 @@ module ButtonsHelper
       data: { confirm: 'All plantings associated with this garden will be marked as finished. Are you sure?' }
   end
 
-  def garden_add_photo_button(garden)
-    button(new_photo_path(id: garden.id, type: "garden"), 'buttons.add_a_photo', 'camera')
-  end
-  
   def seed_edit_button(seed)
     edit_button(edit_seed_path(seed))
   end
-  
+
   def harvest_edit_button(harvest)
     edit_button(edit_harvest_path(harvest))
   end
@@ -42,10 +30,6 @@ module ButtonsHelper
 
   def planting_edit_button(planting)
     edit_button(edit_planting_path(planting))
-  end
-
-  def planting_add_photo_button(planting)
-    button(new_photo_path(id: planting.id, type: 'planting'), 'buttons.add_photo', 'camera')
   end
 
   def planting_finish_button(planting)
@@ -68,17 +52,32 @@ module ButtonsHelper
     button(new_planting_seed_path(planting), 'buttons.save_seeds', 'heart') if can?(:edit, planting)
   end
 
+  # Generic buttons (works out which model)
+  def add_photo_button(model)
+    return unless can?(:edit, model) && can?(:create, Photo)
+
+    button(
+      new_photo_path(id: model.id, type: model_type_for_photo(model)),
+      'buttons.add_a_photo', 'camera')
+  end
+
+
   def edit_button(path)
     button(path, 'buttons.edit', 'pencil')
   end
 
   def delete_button(model, message: 'are_you_sure')
+    return unless can? :destroy, model
     link_to model, method: :delete, data: { confirm: t(message) }, class: 'btn btn-default btn-xs' do
       render 'shared/glyphicon', icon: 'trash', title: 'buttons.delete'
     end
   end
 
   private
+
+  def model_type_for_photo(model)
+    ActiveModel::Name.new(model.class).to_s.downcase
+  end
 
   def button(path, title, icon, size = 'btn-xs')
     link_to path, class: "btn btn-default #{size}" do
